+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Note: Breaking changes between versions are indicated by "💥".
 
 - [Bugfix] Fix insecure static asset loading when web proxy is enabled
 - [Improvement] Rename ``SECRET_KEY`` configuration parameter to ``OPENEDX_SECRET_KEY``
-- [Improvement] Add support for TLS in external SMTP server (#231)
+- [Improvement] Add support for SSL and TLS in external SMTP server (#231)
 - [Bugfix] Fix missing video transcripts in LMS (#229)
 - [Improvement] Make it possible to enable/disable multiple plugins at once
 - [Improvement] Add 'local-docker-compose-nginx-volumes' patch